Create README.md
Simple calculator project to practice Java while in school. While i learn more, features will be added to the calculator like history, extended calculations, and eventually getting rid of the need to ask how the user wants to combine the numbers. Would also like to add a feature to display the output in various ways such as binary.